feat- refactor the file structure for components of article related components
